refactor: implement generalExport function for enhanced data export capabilities
